ai-assistant-backend/pyproject.toml
[tool.poetry]
name = "chat_server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "^2.5.2"
langchain-core = "^0.3.28"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.13"
httpx = "^0.28.1"
langchain-chroma = "^0.1.4"
jq = "^1.8.0"
atlassian-python-api = "^3.41.16"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
